petshop-core-parent/petshop-entity/src/test/java/com/genmymodel/petshop/wrapper/CustomerTest.java
package com.genmymodel.petshop.wrapper;



import org.junit.Assert;
import org.junit.Before;
import org.junit.After;
import org.junit.Test;
import java.util.List;
//Start of user code for imports
// TODO: import me!
//End of user code


/*
 * Test Class
 * <!-- Start of user code comment for Customer Test Class
 * End of user code -->
 */

/**
 *  
 */

public class CustomerTest
{
	//Start of user code for attributes
	// Other attributes
	//End of user code

	private Customer wrapper;	

	@Before
	public void setup(){
		this.wrapper = new Customer();
	}
	
	@After
	public void teardown(){
		this.wrapper = null;
	}

	/*
	 * Basic Atts setter/getter
	 */
	@Test
	public void setAndGetFirstName() {
		//Start of user code basic set and get for firstName attribute
		String tmp = "";
		this.wrapper.setFirstName(tmp);
		String result = this.wrapper.getFirstName();
		Assert.assertNotNull(result);
		Assert.assertEquals(tmp, result);
		//End of user code
	}
	@Test
	public void setAndGetLastName() {
		//Start of user code basic set and get for lastName attribute
		String tmp = "";
		this.wrapper.setLastName(tmp);
		String result = this.wrapper.getLastName();
		Assert.assertNotNull(result);
		Assert.assertEquals(tmp, result);
		//End of user code
	}
	@Test
	public void setAndGetId() {
		//Start of user code basic set and get for id attribute
		Long tmp = 0L;
		this.wrapper.setId(tmp);
		Long result = this.wrapper.getId();
		Assert.assertNotNull(result);
		Assert.assertEquals(tmp, result);
		//End of user code
	}
	@Test
	public void setAndGetCreationDate() {
		//Start of user code basic set and get for creationDate attribute
		java.util.Date tmp = new java.util.Date();
		this.wrapper.setCreationDate(tmp);
		java.util.Date result = this.wrapper.getCreationDate();
		Assert.assertNotNull(result);
		Assert.assertEquals(tmp, result);
		//End of user code
	}
	@Test
	public void setAndGetUpdateDate() {
		//Start of user code basic set and get for updateDate attribute
		java.util.Date tmp = new java.util.Date();
		this.wrapper.setUpdateDate(tmp);
		java.util.Date result = this.wrapper.getUpdateDate();
		Assert.assertNotNull(result);
		Assert.assertEquals(tmp, result);
		//End of user code
	}
	@Test
	public void setAndGetVersion() {
		//Start of user code basic set and get for version attribute
		Integer tmp = 0;
		this.wrapper.setVersion(tmp);
		Integer result = this.wrapper.getVersion();
		Assert.assertNotNull(result);
		Assert.assertEquals(tmp, result);
		//End of user code
	}
	
	/*
	 * Basic Refs setter/getter
	 */
	@Test
	public void setAndGetCarts() {
		//Start of user code basic set and get for carts reference
		Cart tmp = new Cart();
		this.wrapper.setCarts(tmp);
		Cart result = this.wrapper.getCarts();
		Assert.assertNotNull(result);
		Assert.assertEquals(tmp, result);
		//End of user code
	}
	@Test
	public void setAndGetOrders() {
		//Start of user code basic set and get for orders reference
		Order tmp = new Order();
		this.wrapper.getOrders().add(tmp);
		List<Order> result = this.wrapper.getOrders();
		Assert.assertTrue(result.contains(tmp));
		//End of user code
	}
	@Test
	public void setAndGetAddress() {
		//Start of user code basic set and get for address reference
		Address tmp = new Address();
		this.wrapper.setAddress(tmp);
		Address result = this.wrapper.getAddress();
		Assert.assertNotNull(result);
		Assert.assertEquals(tmp, result);
		//End of user code
	}
	

	/*
	 * toXML
	 */
	@Test
	public void toXML() { 
		//Start of user code XML production
		this.wrapper.toXML();
		//End of user code
	}
	
	//Start of user code for test methods
	// Other test methods
	//End of user code
}
